Added logic to support editing and uploading multiple photos in a post
multiple photos can be uploaded with a post, photos larger than 1mb are compressed down to fit the max payload of the firebase call. However the compression algorithm didnt work with gifs so large gifs are uploaded to firebase storage. All imaages are deleted from the db when the post is deleted